Addon-knobs: organize type definitions a bit
